Move translation setup to shared module
