+++ client/app/src/app/shared/shared.module.ts
@@ -8,13 +8,18 @@ import { HttpClientModule, HttpClient } from '@angular/common/http';
 import { Loc } from './_services/location.service';
 import { WindowRef } from './_services/window-ref.service';
 
+import { TranslateModule, TranslateLoader } from '@ngx-translate/core';
+import { TranslateHttpLoader } from '@ngx-translate/http-loader';
+export function HttpClientLoaderFactory(httpClient: HttpClient) {
+  return new TranslateHttpLoader(httpClient, './assets/', '.json');
+}
+
 import {
   RedirectToDefaultRouteComponent,
 } from './_components/redirect-to-default-route/redirect-to-default-route.component';
 import { TangySvgLogoComponent } from './_components/tangy-svg-logo/tangy-svg-logo.component';
 import { TruncateValuePipe } from './_pipes/truncate-value.pipe';
 import { UnsanitizeHtmlPipe } from './_pipes/unsanitize-html.pipe';
-import { TranslateModule } from '@ngx-translate/core';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 @NgModule({
@@ -22,7 +27,14 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
     CommonModule,
     BrowserAnimationsModule,
     HttpClientModule,
-    MatTooltipModule
+    MatTooltipModule,
+    TranslateModule.forRoot({
+      loader: {
+        provide: TranslateLoader,
+        useFactory: HttpClientLoaderFactory,
+        deps: [HttpClient]
+      }
+    })
   ],
   providers: [
     AppConfigService,
